+++ apps/sanity/sanity.config.ts
@@ -3,7 +3,7 @@
 import {defineConfig} from 'sanity'
 import {structureTool} from 'sanity/structure'
 import {visionTool} from '@sanity/vision'
-import {schemaTypes} from './schemaTypes'
+import {schemaTypes} from './schema-types'
 
 export default defineConfig({
   name: 'default',
