+++ apps/site/package.json
@@ -85,7 +85,6 @@
     "stripe": "^20.4.1",
     "tailwind-merge": "^3.5.0",
     "tw-animate-css": "^1.4.0",
-    "vaul": "^1.1.2",
     "zod": "^4.3.6"
   },
   "devDependencies": {
