--- apps/site/src/lib/types/components.ts
@@ -1,13 +1,0 @@
-// Copyright © Todd Agriscience, Inc. All rights reserved.
-
-/** @fileoverview A collection of types relating to components -- generally props. */
-
-/** A menu item for links.
- *
- * @property {string} href - The relative URL
- * @property {string} label - The label on the link itself
- */
-export interface MenuItem {
-  href: string;
-  label: string;
-}
